árfolyam menu hozzáadása; használt árfolyamok számának csökkentése, mértékegységek bővítése; árfolyamválasztó implementálása; originalprice és displayedprice tárolása, megkülönböztetése árfolyamátváltás során előjövő adatkonzisztencia miatt; árfolyamfrissítő opció implementálása; PreferredCurrency szerializálása; PreferredCurrency beállítása új termék és termék szerkesztése ablakokon; CurrencyNumber & StockNumber: float helyett bigdecimal használata; Billentyűzettel való kezelést javítása, ha nem a legelsőt szeretném választani: delCategory requestFocus off, editGoods categoryn requestFocus off, delGoods mind2 requestFocus off; editGoods, delGoods: kiválasztva volt vmi termék majd kiválasztasz egy üres kategóriát!! >> Workaround implemented: törli és letiltja a itemName comboboxot (nem találtam megoldást, valszeg Java11+-ban ilyen, nem lehet visszaállítani a prompt textet); Karakterlimitek bevezetése TextField-ek (nevek, gyártó): 45, TextArea-k (leírás): 200

+++ src/main/java/com/mycompany/raktar/currencyXML/Item.java
@@ -3,14 +3,16 @@ package com.mycompany.raktar.currencyXML;
 import jakarta.xml.bind.annotation.XmlAccessType;
 import jakarta.xml.bind.annotation.XmlAccessorType;
 
+import java.math.BigDecimal;
+
 @XmlAccessorType(XmlAccessType.FIELD)
 public class Item {
     private String penznem;
-    private Float kozep;
+    private BigDecimal kozep;
 
     public Item() {}
 
-    public Item(String penznem, Float kozep) {
+    public Item(String penznem, BigDecimal kozep) {
         super();
         this.penznem = penznem;
         this.kozep = kozep;
@@ -20,7 +22,7 @@ public class Item {
         return penznem;
     }
 
-    public Float getKozep() {
+    public BigDecimal getKozep() {
         return kozep;
     }
 }
